infra/infra_manual/auxiliar_files/functions/deploy_notification/requirements.txt
python-telegram-bot==21.4
# python-telegram-bot==20.2